% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame describing the graph. Must contain three variables (v1, v2, and w) containing the edges of the graph (from v1 to v2) and the weight of the edge (w)}

\item{init_node}{A numeric scalar representing the initial node from which to calculate shortest paths}
}
\value{
A numeric vector of shortest distances from init_node to all other nodes
}
\description{
Calculate the shortest path from an initial node to all other nodes in the graph
using Dijkstra's algorithm. The algorithm maintains a set of unvisited nodes and
iteratively selects the unvisited node with the smallest distance, updating the
distances to its neighbors until all nodes have been processed.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)

}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
