% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The greatest common divisor of a and b
}
\description{
Find the greatest common divisor of two numbers using the Euclidean algorithm.
The algorithm repeatedly applies the division algorithm until the remainder is zero.
The last non-zero remainder is the greatest common divisor.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
