% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia example graph}
\format{
A data frame with 18 rows and 3 columns:
\describe{
  \item{v1}{Vertex 1 - from where the edge is outgoing}
  \item{v2}{Vertex 2 - where the edge is incoming}
  \item{w}{The weight of the edge}
}
}
\source{
<https://en.wikipedia.org/wiki/Euclidean_algorithm>
}
\usage{
wiki_graph
}
\description{
A dataset describing a graph to use as example for Dijkstra's algorithm
}
\keyword{datasets}
