% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The greatest common divisor of a and b
}
\description{
Find the greatest common divisor of two numbers
TODO description of algorithm
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
